get a minimal websocket example running on ddev with racket
