+++ bin/chat-server.php
@@ -1,41 +1,18 @@
 <?php
-namespace MyApp;
 use Ratchet\Server\IoServer;
-use MyApp\Chat;
-
-use Ratchet\MessageComponentInterface;
-use Ratchet\ConnectionInterface;
-require dirname(__DIR__) . '/vendor/autoload.php';
-
-
-class Chat implements MessageComponentInterface {
-    public function onOpen(ConnectionInterface $conn) {
-    }
-
-    public function onMessage(ConnectionInterface $from, $msg) {
-    }
-
-    public function onClose(ConnectionInterface $conn) {
-    }
-
-    public function onError(ConnectionInterface $conn, \Exception $e) {
-    }
-}
-
-
-
 use Ratchet\Http\HttpServer;
 use Ratchet\WebSocket\WsServer;
+use MyApp\Chat;
 
-    require dirname(__DIR__) . '/vendor/autoload.php';
+require dirname(__DIR__) . '/vendor/autoload.php';
 
-    $server = IoServer::factory(
-        new HttpServer(
-            new WsServer(
-                new Chat()
-            )
-        ),
-        8080
-    );
+$server = IoServer::factory(
+    new HttpServer(
+        new WsServer(
+            new Chat()
+        )
+    ),
+    8080
+);
 
-    $server->run();
+$server->run();
